Split out s3m playback objects
