+++ internal/format/s3m/playback/effect/effect_stereocontrol.go
@@ -3,8 +3,8 @@ package effect
 import (
 	"fmt"
 
+	s3mPanning "github.com/gotracker/gotracker/internal/format/s3m/conversion/panning"
 	"github.com/gotracker/gotracker/internal/format/s3m/layout/channel"
-	"github.com/gotracker/gotracker/internal/format/s3m/playback/util"
 	"github.com/gotracker/gotracker/internal/player/intf"
 )
 
@@ -18,9 +18,9 @@ func (e StereoControl) Start(cs intf.Channel[channel.Memory, channel.Data], p in
 	x := uint8(e) & 0xf
 
 	if x > 7 {
-		cs.SetPan(util.PanningFromS3M(x - 8))
+		cs.SetPan(s3mPanning.PanningFromS3M(x - 8))
 	} else {
-		cs.SetPan(util.PanningFromS3M(x + 8))
+		cs.SetPan(s3mPanning.PanningFromS3M(x + 8))
 	}
 	return nil
 }
